Fix/controller tests

+++ deliveries/src/main/java/com/meds/deliveries/service/PersonService.java
@@ -2,72 +2,81 @@ package com.meds.deliveries.service;
 
 import com.meds.deliveries.model.Person;
 import com.meds.deliveries.repository.PersonRepository;
+
+import lombok.NonNull;
+
 import com.meds.deliveries.exception.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @Service
 public class PersonService {
-    @Autowired
-    private PersonRepository repository;
+    
+    @Autowired PersonRepository repository;
 
-    public Person savePerson(Person Person) {
-        return repository.save(Person);
-    }
-
-    public List<Person> savePersons(List<Person> Persons) {
-        return repository.saveAll(Persons);
-    }
-
-    public List<Person> getPersons() {
-        return repository.findAll();
+    public void registerPerson(Person p) throws DuplicatedObjectException {
+        if (repository.existsByUsername(p.getUsername()))
+            throw new DuplicatedObjectException("The provided username is already taken.");
+        repository.save(p);    
     }
 
     public Person getPersonById(int id) throws ResourceNotFoundException {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Person not found for this id:" + id);
+        }
+        return (Person) repository.findById(id).get();
 
-        Person p = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Person not found for this username:" + id));
-        return p;
-        
     }
-    
+
+    @NonNull
     public Person getPersonByUsername(String username) {
-        return repository.findByUsername(username).orElseThrow(() -> new ResourceNotFoundException("Person not found for this username:" + username));
+        if (!repository.existsByUsername(username)) throw new ResourceNotFoundException("Person not found for this username:" + username);
+        
+        return (Person) repository.findByUsername(username).get();
     }
 
+    @NonNull
     public Person getPersonByEmail(String email) {
-        return repository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("Person not found for this email:" + email));
+        if (!repository.existsByEmail(email)) {
+            throw new ResourceNotFoundException("Person not found for this email:" + email);
+        }
+        return (Person) repository.findByEmail(email).get();
     }
-  
+
     public Map<String, Boolean> deletePerson(Person Person) throws ResourceNotFoundException {
 
-        int id  = Person.getId();
+        int id = Person.getId();
 
-        repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Person not found for this id:" + id));
-    
+        if (!repository.existsById(id)) {
+            
+            throw new ResourceNotFoundException("Person not found for this id:" + id);
+        }
+
         repository.deleteById(id);
         Map<String, Boolean> response = new HashMap<>();
         response.put("deleted", Boolean.TRUE);
-        return response;  
+        return response;
     }
 
     public Person updatePerson(Person person) throws ResourceNotFoundException {
         int id = person.getId();
-        Person existingPerson = repository.findById(id)
-        .orElseThrow(() -> new ResourceNotFoundException("Person not found for this id:" + id));
+        if (!repository.existsById(id)) {
+            System.out.println("cant found");
+            throw new ResourceNotFoundException("Person not found for this id:" + id);
+        }
+
+        Person existingPerson = repository.findById(id).get();
 
         existingPerson.setName(person.getName());
         existingPerson.setEmail(person.getEmail());
         existingPerson.setPhone(person.getPhone());
         existingPerson.setUsername(person.getUsername());
         existingPerson.setPassword(person.getPassword());
-
 
         return repository.save(existingPerson);
     }
